requirements.txt
Babel==1.3
boto==2.38.0
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.0
flask-oidc==0.1.2
Flask-OpenID==1.2.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
Whoosh==2.6.0
blinker==1.3
coverage==3.7.1
decorator==3.4.0
flipflop==1.0
gunicorn==19.1.1
itsdangerous==0.24
pbr==0.10.0
Pillow==2.8.1
python-slugify==0.1.0
psycopg2==2.5.4
python-openid==2.2.5
pytz==2014.7
six==1.8.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlparse==0.1.11
rauth==0.7.1


